+++ src/tkinteros/callback_management/callback_manager.py
