+++ src/tkinteros/gui/taskbar_gui.py
@@ -1,6 +1,6 @@
 from PIL import Image
 from typing import Callable
-from desktop.callbacks import Callback
+from callback_management.callbacks import Callback
 
 import customtkinter as ctk
