+++ src/tkinteros/gui/text_editor.py
